+++ src/cpp/rtps/messages/submessages/DataMsg.hpp
@@ -204,7 +204,7 @@ bool RTPSMessageCreator::addSubmessageData(
 
 
         //TODO(Ricardo) Improve.
-        submessage_size = msg->pos - position_size_count_size;
+        submessage_size = uint16_t(msg->pos - position_size_count_size);
         octet* o= (octet*)&submessage_size;
         if(msg->msg_endian == DEFAULT_ENDIAN)
         {
@@ -430,7 +430,7 @@ bool RTPSMessageCreator::addSubmessageDataFrag(
             added_no_error &= CDRMessage::addOctet(msg, 0);
 
         //TODO(Ricardo) Improve.
-        submessage_size = msg->pos - position_size_count_size;
+        submessage_size = uint16_t(msg->pos - position_size_count_size);
         octet* o= (octet*)&submessage_size;
         if(msg->msg_endian == DEFAULT_ENDIAN)
         {
